+++ src/lmgrep/unordered.clj
@@ -1,22 +1,9 @@
 (ns lmgrep.unordered
-  (:require [lmgrep.matching :as matching])
+  (:require [lmgrep.concurrent :as c]
+            [lmgrep.matching :as matching])
   (:import (java.io BufferedReader PrintWriter BufferedWriter FileReader)
-           (java.util.concurrent ThreadPoolExecutor TimeUnit Executors LinkedBlockingQueue
-                                 ThreadPoolExecutor$CallerRunsPolicy ExecutorService)
+           (java.util.concurrent ExecutorService)
            (lmgrep.matching LineNrStr)))
-
-(defn shutdown-thread-pool-executors [& executors]
-  (doseq [^ExecutorService executor executors]
-    (.shutdown executor)
-    (.awaitTermination executor 60 TimeUnit/SECONDS)))
-
-(defn thread-pool-executor [^Integer concurrency ^Integer queue-size]
-  (ThreadPoolExecutor.
-    concurrency concurrency
-    0 TimeUnit/MILLISECONDS
-    (LinkedBlockingQueue. queue-size)
-    (Executors/defaultThreadFactory)
-    (ThreadPoolExecutor$CallerRunsPolicy.)))
 
 (defn consume-reader
   "Given a Reader iterates over lines and sends them to the
@@ -48,8 +35,8 @@
         queue-size (get options :queue-size 1024)
         with-empty-lines (get options :with-empty-lines)
         ^PrintWriter writer (PrintWriter. (BufferedWriter. *out* print-writer-buffer-size))
-        ^ExecutorService matcher-thread-pool-executor (thread-pool-executor concurrency queue-size)
-        ^ExecutorService writer-thread-pool-executor (Executors/newSingleThreadExecutor)]
+        ^ExecutorService matcher-thread-pool-executor (c/thread-pool-executor concurrency queue-size)
+        ^ExecutorService writer-thread-pool-executor (c/single-thread-executor)]
     (doseq [^String path (if (empty? file-paths-to-analyze)
                            [nil]                            ;; STDIN is an input
                            file-paths-to-analyze)]
@@ -62,5 +49,5 @@
                                                               writer
                                                               with-empty-lines)]
         (consume-reader reader unordered-matcher-fn matcher-thread-pool-executor)))
-    (shutdown-thread-pool-executors matcher-thread-pool-executor writer-thread-pool-executor)
+    (c/shutdown-thread-pool-executors matcher-thread-pool-executor writer-thread-pool-executor)
     (.flush writer)))